@@ -11,6 +11,7 @@
       :aria-expanded="!!isHovered || !!isFocused"
       aria-haspopup="true"
       :class="['vts-dropdown__trigger', classes.trigger]"
+      type="button"
       @click="isFocused = !isFocused"
     >
       <!-- @slot The content within the trigger button -->
@@ -54,6 +55,7 @@ export default {
     position: {
       type: String,
       default: 'bottom',
+      /** @param {string} value */
       validator(value) {
         return ['top', 'bottom'].includes(value);
       },
@@ -78,11 +80,15 @@ export default {
   }),
 
   mounted() {
-    const { onClickout } = this;
-    document.addEventListener('click', onClickout);
-    this.$once('hook:beforeDestroy', () => {
-      document.removeEventListener('click', onClickout);
-    });
+    document.addEventListener('click', this.onClickout);
+  },
+
+  /** @deprecated in Vue 3 */
+  beforeDestroy() {
+    document.removeEventListener('click', this.onClickout);
+  },
+  beforeUnmount() {
+    document.removeEventListener('click', this.onClickout);
   },
 
   methods: {
@@ -110,11 +116,11 @@ export default {
 .vts-dropdown__content {
   position: absolute;
   z-index: 5;
-  min-width: 100%;
+  min-inline-size: 100%;
 }
 
 .vts-dropdown__content--top {
-  top: 0;
+  inset-block-start: 0;
   transform: translateY(-100%);
 }
 </style>
