@@ -1,5 +1,5 @@
 <script>
-import { safeSlot } from '../../utils';
+import { isVue3 } from 'vue-demi';
 
 /**
  * Uses [IntersectionObserver](https://developer.mozilla.org/en-US/docs/Web/API/IntersectionObserver) to fire events when content enters or exits the screen.
@@ -34,12 +34,18 @@ export default {
       default: () => ({}),
     },
   },
+  emits: ['enter', 'exit', 'change'],
 
   data: () => ({
     entry: {},
   }),
 
   mounted() {
+    let el = this.$el;
+    if (isVue3) {
+      // I'm not really sure why this is necessary
+      el = this.$el.nextElementSibling;
+    }
     const { root, threshold, rootMargin, options, handler } = this;
     const observerOptions = {
       ...options,
@@ -49,11 +55,14 @@ export default {
     };
 
     this.observer = new IntersectionObserver(handler, observerOptions);
-    this.observer.observe(this.$el);
-
-    this.$once('hook:beforeDestroy', () => {
-      this.observer.disconnect();
-    });
+    this.observer.observe(el);
+  },
+  beforeUnmount() {
+    this.observer.disconnect();
+  },
+  /** @deprecated */
+  beforeDestroy() {
+    this.observer.disconnect();
   },
 
   methods: {
@@ -87,19 +96,22 @@ export default {
     },
   },
 
-  render(h) {
+  render() {
     /** @slot Content to be tracked with IntersectionObserver */
     const { entry } = this;
+    if (isVue3) {
+      return this.$slots.default(entry);
+    }
 
     /** @slot Slot content providing isIntersecting */
     const defaultSlot = this.$slots.default;
     const scopedSlot = this.$scopedSlots.default;
 
     if (defaultSlot) {
-      return safeSlot(h, defaultSlot);
+      return defaultSlot;
     }
 
-    return safeSlot(h, scopedSlot(entry));
+    return scopedSlot(entry);
   },
 };
 </script>
