@@ -2,7 +2,7 @@
   <transition :name="transition">
     <component
       :is="tag"
-      v-if="!dismissed && !!visible"
+      v-if="!dismissed && !!modelValue"
       role="alert"
       :class="['vts-alert', classes.root]"
     >
@@ -13,12 +13,11 @@
         v-if="dismissible"
         :aria-label="dismissLabel"
         :class="['vts-alert__dismiss', classes.dismiss]"
+        type="button"
         @click="dismiss"
       >
         <!-- @slot The dismiss button content -->
-        <slot name="dismiss">
-          &times;
-        </slot>
+        <slot name="dismiss">&times;</slot>
       </button>
     </component>
   </transition>
@@ -26,13 +25,25 @@
 
 <script>
 /**
+ * NOTES Concerning :visible prop.
+ * Per W3C specifications:
+ * https://www.w3.org/TR/wai-aria-practices/#alert
+ * --
+ * "It is also important to avoid designing alerts that disappear automatically.
+ * An alert that disappears too quickly can lead to failure to meet WCAG 2.0 success criterion 2.2.3.
+ * Another critical design consideration is the frequency of interruption caused by alerts.
+ * Frequent interruptions inhibit usability for people with visual and cognitive disabilities,
+ * which makes meeting the requirements of WCAG 2.0 success criterion 2.2.4 more difficult."
+ */
+
+/**
  * A simple component for notifiying users of specific information. Good for informative snippets, error messages, and more. It can be shown or hidden dynamically, and even supports auto-hiding after a given time.
  */
 export default {
   name: 'VAlert',
   model: {
-    prop: 'visible',
-    event: 'update',
+    prop: 'modelValue',
+    event: 'update:modelValue',
   },
 
   props: {
@@ -46,7 +57,7 @@ export default {
     /**
      * Determines whether the alert is visible. Also binds with `v-model`.
      */
-    visible: {
+    modelValue: {
       type: [Boolean, Number],
       default: true,
     },
@@ -81,7 +92,7 @@ export default {
   }),
 
   watch: {
-    visible: {
+    modelValue: {
       handler(visible) {
         if (visible) {
           this.dismissed = false;
@@ -95,6 +106,10 @@ export default {
     },
   },
 
+  beforeUnmount() {
+    this.clearTimer();
+  },
+  /** @deprecated */
   beforeDestroy() {
     this.clearTimer();
   },
@@ -109,18 +124,21 @@ export default {
        */
       this.$emit('dismiss');
       this.dismissed = true;
-      if (typeof this.visible === 'number') {
+      if (typeof this.modelValue === 'number') {
+        this.$emit('update:modelValue', 0);
+        /** @deprecated */
         this.$emit('update', 0);
         this.clearTimer();
       } else {
+        this.$emit('update:modelValue', false);
+        /** @deprecated */
         this.$emit('update', false);
       }
     },
 
     countdown() {
-      const { visible } = this;
-      if (visible <= 0) return;
-
+      const { modelValue } = this;
+      if (typeof modelValue !== 'number' || modelValue <= 0) return;
       this.timerId = setTimeout(() => {
         /**
          * Fired whenever the visibility changes. Either through user interaction, or a countdown timer
@@ -128,7 +146,9 @@ export default {
          * @event update
          * @type { boolean|number }
          */
-        this.$emit('update', visible - 1);
+        this.$emit('update:modelValue', modelValue - 1);
+        /** @deprecated */
+        this.$emit('update', modelValue - 1);
       }, 1000);
     },
 
