@@ -1,8 +1,51 @@
+<template>
+  <span v-if="localShow || slots.toggle">
+    <slot
+      v-if="slots.toggle"
+      name="toggle"
+      v-bind="{
+        on: {
+          click: () => (localShow = !localShow),
+        },
+        bind: {
+          type: 'button',
+          role: 'button',
+          'aria-haspopup': true,
+          'aria-expanded': '' + localShow,
+        },
+      }"
+    />
+    <transition :name="bgTransition || transition" :appear="true">
+      <template v-if="localShow">
+        <div :class="['vts-dialog', classes.root, classes.bg, $attrs.class]">
+          <transition :name="contentTransition">
+            <component
+              :is="tag"
+              ref="content"
+              :class="['vts-dialog__content', classes.content]"
+              :style="{
+                width: width,
+                'inline-size': inlineSize,
+                'max-width': maxWidth,
+                'max-inline-size': maxInlineSize,
+              }"
+              tabindex="-1"
+              role="dialog"
+              aria-modal="true"
+            >
+              <slot v-bind="{ close: () => (localShow = !localShow) }" />
+            </component>
+          </transition>
+        </div>
+      </template>
+    </transition>
+  </span>
+</template>
 <script>
-import KEYCODES from '../../data/keycodes';
-import FOCUSABLE from '../../data/focusable';
+import { isVue3 } from 'vue-demi';
+import KEYCODES from '../../data/keycodes.js';
+import FOCUSABLE from '../../data/focusable.js';
 
-const NAME = 'vts-dialog';
 /**
  * A dialog component for showing users content which overlays the rest of the applications. When opened, it traps the user's focus so that keyboard navigation will remain within the dialog until it is closed. It supports being closed by clicking outside the dialog content or pressing the ESC key.
  */
@@ -12,10 +55,11 @@ export default {
 
   model: {
     prop: 'showing',
-    event: 'update',
+    event: 'update:modelValue',
   },
 
   props: {
+    modelValue: Boolean,
     /**
      * @model
      */
@@ -42,9 +86,23 @@ export default {
       default: '',
     },
     /**
+     * CSS width to set the dialog to.
+     */
+    inlineSize: {
+      type: String,
+      default: '',
+    },
+    /**
      * CSS max-width to set the dialog to.
      */
     maxWidth: {
+      type: String,
+      default: '',
+    },
+    /**
+     * CSS max-width to set the dialog to.
+     */
+    maxInlineSize: {
       type: String,
       default: '',
     },
@@ -61,8 +119,18 @@ export default {
     },
     /**
      * Transition name to apply to the background.
+     *
+     * @deprecated
      */
     bgTransition: {
+      type: String,
+      default: '',
+    },
+
+    /**
+     * Transition name to apply to the background.
+     */
+    contentTransition: {
       type: String,
       default: '',
     },
@@ -72,16 +140,29 @@ export default {
       default: () => ({}),
     },
   },
+  emits: ['update', 'update:modelValue', 'open', 'close'],
 
   data() {
     return {
-      localShow: this.showing,
+      localShow: this.modelValue || this.showing,
       activeElement: null,
     };
   },
 
+  computed: {
+    slots() {
+      if (!isVue3) {
+        return this.$scopedSlots;
+      }
+      return this.$slots;
+    },
+  },
+
   watch: {
     showing(next) {
+      this.localShow = next;
+    },
+    modelValue(next) {
       this.localShow = next;
     },
     localShow: {
@@ -103,6 +184,7 @@ export default {
         }
 
         this.$emit('update', next);
+        this.$emit('update:modelValue', next);
       },
     },
   },
@@ -118,12 +200,14 @@ export default {
       window.addEventListener('keydown', onKeydown);
       noScroll && document.body.style.setProperty('overflow', 'hidden');
       this.$nextTick(() => this.$refs.content.focus());
+      this.$emit('open');
     },
     onClose() {
       const { onClick, onKeydown, noScroll } = this;
       window.removeEventListener('click', onClick);
       window.removeEventListener('keydown', onKeydown);
       noScroll && document.body.style.removeProperty('overflow');
+      this.$emit('close');
     },
     onClick(event) {
       if (event.target.classList.contains('vts-dialog') && this.dismissible) {
@@ -135,6 +219,7 @@ export default {
         this.localShow = false;
       }
       if (event.keyCode === KEYCODES.TAB) {
+        /** @type {HTMLDivElement} */
         const content = this.$refs.content;
         if (!content) return;
 
@@ -164,88 +249,6 @@ export default {
       }
     },
   },
-
-  render(h) {
-    const { localShow, $scopedSlots, classes } = this;
-
-    if (!localShow && !$scopedSlots.toggle) {
-      return h(false);
-    }
-
-    const children = [];
-
-    if ($scopedSlots.toggle) {
-      children.push(
-        $scopedSlots.toggle({
-          on: {
-            click: () => (this.localShow = true),
-          },
-          bind: {
-            type: 'button',
-            role: 'button',
-            'aria-haspopup': true,
-            'aria-expanded': '' + localShow,
-          },
-          attrs: {
-            // TODO: deprecated
-            type: 'button',
-            role: 'button',
-            'aria-haspopup': true,
-            'aria-expanded': '' + localShow,
-          },
-        })
-      );
-    }
-
-    if (localShow) {
-      let content = h(
-        this.tag,
-        {
-          ref: 'content',
-          class: [`${NAME}__content`, classes.content],
-          style: {
-            width: this.width,
-            maxWidth: this.maxWidth,
-          },
-          attrs: {
-            tabindex: '-1',
-            role: 'dialog',
-          },
-        },
-        [this.$slots.default]
-      );
-      content = h(
-        'transition',
-        {
-          props: { name: this.transition },
-        },
-        [content]
-      );
-
-      const modal = h(
-        'div',
-        {
-          class: [NAME, classes.root, classes.bg, this.$attrs.class],
-        },
-        [content]
-      );
-
-      children.push(
-        h(
-          'transition',
-          {
-            props: {
-              name: this.bgTransition,
-              appear: true
-            },
-          },
-          [modal]
-        )
-      );
-    }
-
-    return h('span', children);
-  },
 };
 </script>
 
@@ -256,10 +259,7 @@ export default {
   justify-content: center;
   position: fixed;
   z-index: 100;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
+  inset: 0;
 }
 
 .vts-dialog__content:focus {
