@@ -1,6 +1,55 @@
+<template>
+  <span v-if="localShow || slots.toggle">
+    <slot
+      v-if="slots.toggle"
+      name="toggle"
+      v-bind="{
+        on: {
+          click: () => (localShow = !localShow),
+        },
+        bind: {
+          type: 'button',
+          role: 'button',
+          'aria-haspopup': true,
+          'aria-expanded': '' + localShow,
+        },
+      }"
+    />
+    <transition :name="bgTransition || transition" appear>
+      <template v-if="localShow">
+        <div :class="['vts-drawer', classes.root, classes.bg, $attrs.class]">
+          <transition :name="transition" appear>
+            <component
+              :is="tag"
+              v-if="localShow"
+              ref="content"
+              :class="[
+                'vts-drawer__content',
+                { 'vts-drawer__content--right': !!right },
+                classes.content,
+              ]"
+              :style="{
+                width: width,
+                'inline-size': inlineSize,
+                'max-width': maxWidth,
+                'max-inline-size': maxInlineSize,
+              }"
+              tabindex="-1"
+            >
+              <!-- role="dialog" TODO ?? -->
+              <slot v-bind="{ close: () => (localShow = !localShow) }" />
+            </component>
+          </transition>
+        </div>
+      </template>
+    </transition>
+  </span>
+</template>
+
 <script>
-import KEYCODES from '../../data/keycodes';
-import FOCUSABLE from '../../data/focusable';
+import { isVue3 } from 'vue-demi';
+import KEYCODES from '../../data/keycodes.js';
+import FOCUSABLE from '../../data/focusable.js';
 
 const NAME = 'vts-drawer';
 
@@ -11,10 +60,11 @@ export default {
   name: 'VDrawer',
   model: {
     prop: 'showing',
-    event: 'update',
+    event: 'update:modelValue',
   },
 
   props: {
+    modelValue: Boolean,
     /**
      * @model
      */
@@ -35,9 +85,23 @@ export default {
       default: '',
     },
     /**
+     * CSS width to set the dialog to.
+     */
+    inlineSize: {
+      type: String,
+      default: '',
+    },
+    /**
      * CSS max-width value.
      */
     maxWidth: {
+      type: String,
+      default: '',
+    },
+    /**
+     * CSS max-width to set the dialog to.
+     */
+    maxInlineSize: {
       type: String,
       default: '',
     },
@@ -54,6 +118,8 @@ export default {
     },
     /**
      * Vue transition name for the background.
+     *
+     * @deprecated
      */
     bgTransition: {
       type: String,
@@ -65,16 +131,29 @@ export default {
       default: () => ({}),
     },
   },
+  emits: ['update', 'update:modelValue', 'open', 'close'],
 
   data() {
     return {
-      localShow: this.showing,
+      localShow: this.modelValue || this.showing,
       activeElement: null,
     };
   },
 
+  computed: {
+    slots() {
+      if (!isVue3) {
+        return this.$scopedSlots;
+      }
+      return this.$slots;
+    },
+  },
+
   watch: {
     showing(next) {
+      this.localShow = next;
+    },
+    modelValue(next) {
       this.localShow = next;
     },
     localShow: {
@@ -96,6 +175,7 @@ export default {
         }
 
         this.$emit('update', next);
+        this.$emit('update:modelValue', next);
       },
     },
   },
@@ -111,12 +191,14 @@ export default {
       window.addEventListener('keydown', onKeydown);
       noScroll && document.body.style.setProperty('overflow', 'hidden');
       this.$nextTick(() => this.$refs.content.focus());
+      this.$emit('open');
     },
     onClose() {
       const { onClick, onKeydown, noScroll } = this;
       window.removeEventListener('click', onClick);
       window.removeEventListener('keydown', onKeydown);
       noScroll && document.body.style.removeProperty('overflow');
+      this.$emit('close');
     },
     onClick(event) {
       if (event.target.classList.contains(NAME)) {
@@ -128,6 +210,7 @@ export default {
         this.localShow = false;
       }
       if (event.keyCode === KEYCODES.TAB) {
+        /** @type {HTMLDivElement} */
         const content = this.$refs.content;
         if (!content) return;
 
@@ -157,95 +240,6 @@ export default {
       }
     },
   },
-
-  render(h) {
-    const { localShow, $scopedSlots, classes } = this;
-
-    if (!localShow && !$scopedSlots.toggle) {
-      return h(false);
-    }
-
-    const children = [];
-
-    if ($scopedSlots.toggle) {
-      children.push(
-        $scopedSlots.toggle({
-          on: {
-            click: () => (this.localShow = true),
-          },
-          bind: {
-            type: 'button',
-            role: 'button',
-            'aria-haspopup': true,
-            'aria-expanded': '' + localShow,
-          },
-          attrs: {
-            // TODO: deprecated
-            type: 'button',
-            role: 'button',
-            'aria-haspopup': true,
-            'aria-expanded': '' + localShow,
-          },
-        })
-      );
-    }
-
-    if (localShow) {
-      let content = h(
-        this.tag,
-        {
-          ref: 'content',
-          class: [
-            `${NAME}__content`,
-            { 'vts-drawer__content--right': !!this.right },
-            classes.content,
-          ],
-          style: {
-            width: this.width,
-            maxWidth: this.maxWidth,
-          },
-          attrs: {
-            tabindex: '-1',
-            // role: "dialog", // TODO: ?
-          },
-        },
-        [this.$slots.default]
-      );
-      content = h(
-        'transition',
-        {
-          props: {
-            name: this.transition,
-            appear: true 
-          },
-        },
-        [content]
-      );
-
-      const component = h(
-        'div',
-        {
-          class: [NAME, classes.root, classes.bg, this.$attrs.class],
-        },
-        [content]
-      );
-
-      children.push(
-        h(
-          'transition',
-          {
-            props: {
-              name: this.bgTransition,
-              appear: true 
-            },
-          },
-          [component]
-        )
-      );
-    }
-
-    return h('span', children);
-  },
 };
 </script>
 
@@ -253,16 +247,13 @@ export default {
 .vts-drawer {
   position: fixed;
   z-index: 100;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
+  inset: 0;
 }
 
 .vts-drawer__content {
   overflow: auto;
-  max-width: 300px;
-  height: 100%;
+  max-inline-size: 20rem;
+  block-size: 100%;
 }
 
 .vts-drawer__content:focus {
@@ -270,6 +261,6 @@ export default {
 }
 
 .vts-drawer__content--right {
-  margin-left: auto;
+  margin-inline-start: auto;
 }
 </style>
