@@ -1,5 +1,5 @@
 <template>
-  <div v-if="tablist.length" :class="['vts-tabs', classes.root]">
+  <div :class="['vts-tabs', classes.root]">
     <div
       role="tablist"
       :aria-label="label"
@@ -7,29 +7,44 @@
       :class="['vts-tabs__tablist', classes.tablist]"
     >
       <button
-        v-for="(tab, index) in tablist"
-        :id="`${id}-tab-${index}`"
+        v-for="(tab, index) in tabList"
+        :id="`${id}-${tab}`"
         :key="tab"
-        ref="tab"
+        :ref="tab"
         :aria-selected="index === activeIndex"
-        :tabindex="index === activeIndex ? false : -1"
-        :aria-controls="`${id}-panel-${index}`"
-        :class="[`vts-tabs__tab vts-tabs__tab--${index}`, classes.tab]"
+        :tabindex="index === activeIndex ? null : -1"
+        :aria-controls="`${id}-${tab.replace('tab', 'panel')}`"
+        :class="[
+          `vts-tabs__tab vts-tabs__tab--${tab} vts-tabs__tab--${index}`,
+          {
+            'vts-tabs__tab--active': index === activeIndex,
+            [classes.tabActive]: index === activeIndex,
+          },
+          classes.tab,
+        ]"
         role="tab"
+        type="button"
         @keydown="onKeydown($event, tab, index)"
         @click="onClick($event, tab, index)"
       >
-        {{ tab }}
+        <slot :name="tab" />
       </button>
     </div>
 
     <div
-      v-for="(tab, index) in tablist"
+      v-for="(tab, index) in panelList"
       :id="`${id}-panel-${index}`"
       :key="tab"
       :aria-labelledby="`${id}-tab-${index}`"
       :hidden="index !== activeIndex"
-      :class="[`vts-tabs__panel vts-tabs__panel--${index}`, classes.panel]"
+      :class="[
+        `vts-tabs__panel vts-tabs__panel--${index}`,
+        {
+          'vts-tabs__panel--active': index === activeIndex,
+          [classes.panelActive]: index === activeIndex,
+        },
+        classes.panel,
+      ]"
       tabindex="0"
       role="tabpanel"
     >
@@ -39,10 +54,12 @@
 </template>
 
 <script>
-import { randomString } from '../../utils';
-import keycodes from '../../data/keycodes';
+import { randomString } from '../../utils.js';
+import keycodes from '../../data/keycodes.js';
 
 // const NAME = "vts-tabs"
+
+// https://codesandbox.io/embed/vue-tabs-pt5lm?codemirror=1
 
 /**
  * Show and hide content based on which tabs are selected.
@@ -54,17 +71,20 @@ import keycodes from '../../data/keycodes';
 export default {
   name: 'VTabs',
 
+  model: {
+    prop: 'active',
+    event: 'change',
+  },
+
   props: {
-    /**
-     * Support for aria-label attribute
-     */
+    active: {
+      type: Number,
+      default: 0,
+    },
     label: {
       type: String,
       default: undefined,
     },
-    /**
-     * Support for aria-orientation attribute
-     */
     orientation: {
       type: String,
       default: 'horizontal',
@@ -76,19 +96,37 @@ export default {
     },
   },
 
-  data: () => ({
-    activeIndex: 0,
-  }),
+  data() {
+    return {
+      activeIndex: this.active,
+      activeTab: this.active,
+    };
+  },
 
   computed: {
-    tablist() {
-      return Object.keys(this.$slots);
+    /** @returns {Array} */
+    tabList() {
+      return Object.keys(this.$slots).filter((name) => name.startsWith('tab-'));
+    },
+    /** @returns {Array} */
+    panelList() {
+      return Object.keys(this.$slots).filter((name) =>
+        name.startsWith('panel-')
+      );
+    },
+  },
+
+  watch: {
+    active(next) {
+      this.activeIndex = Math.max(0, Math.min(this.tabList.length - 1, next));
+    },
+    activeIndex(next) {
+      this.$emit('change', next);
     },
   },
 
   created() {
-    const { id } = this.$attrs;
-    this.id = id ? id : `vts-${randomString(4)}`;
+    this.id = this.$attrs.id || `vts-${randomString(4)}`;
   },
 
   methods: {
@@ -104,7 +142,7 @@ export default {
       switch (keyCode) {
         case keycodes.END:
           event.preventDefault();
-          this.activeIndex = this.tablist.length - 1;
+          this.activeIndex = this.tabList.length - 1;
           this.setFocus();
           break;
         case keycodes.HOME:
@@ -130,7 +168,7 @@ export default {
       }
     },
 
-    // When a tablist's aria-orientation is set to vertical, only up and down arrow should function. In all other cases only left and right arrow function.
+    // When a tabList's aria-orientation is set to vertical, only up and down arrow should function. In all other cases only left and right arrow function.
     determineOrientation(event) {
       const keyCode = event.keyCode;
       let proceed = false;
@@ -141,6 +179,7 @@ export default {
         }
       } else {
         if (keyCode === keycodes.LEFT || keyCode === keycodes.RIGHT) {
+          event.preventDefault();
           proceed = true;
         }
       }
@@ -164,7 +203,7 @@ export default {
       if (!directions[keyCode]) return;
 
       const activeIndex = this.activeIndex;
-      const tabLength = this.$refs.tab.length;
+      const tabLength = this.tabList.length;
       const nextIndex = activeIndex + directions[keyCode];
 
       if (nextIndex < 0) {
@@ -177,7 +216,10 @@ export default {
     },
 
     setFocus() {
-      this.$refs.tab[this.activeIndex].focus();
+      const { $refs, tabList, activeIndex } = this;
+      const activeTab = tabList[activeIndex];
+
+      return $refs[activeTab][0].focus();
     },
   },
 };
