@@ -4,9 +4,9 @@ import FOCUSABLE from './data/focusable.js';
  * Generates a random string of defined length based on
  * a string of allowed characters.
  *
- * @param  {number} length  How many random characters will be in the returned string. Defaults to 10
- * @param  {string} allowed Which characters can be used when creating the random string. Defaults to A-Z,a-z,0-9
- * @return {string}         A string of random characters
+ * @param  {number} [length]  How many random characters will be in the returned string. Defaults to 10
+ * @param  {string} [allowed] Which characters can be used when creating the random string. Defaults to A-Z,a-z,0-9
+ * @returns {string}         A string of random characters
  */
 export function randomString(
   length = 10,
@@ -20,14 +20,17 @@ export function randomString(
 }
 
 /**
- * If multiple elements are passed into a slot, this wraps them in a div
+ * Tells you if a given value matches the type you pass.
  *
- * @param  {Function} h     hyperscript markup from Vue render function
- * @param  {Array}    slot  Array of VDOM nodes passed into component slot
- * @return {Array}          VDOM node of the original Slot content or it wrapped in a div
+ * @param {any} v The value to get the type of
+ * @param {string} type The type you are asserting against
+ * @returns {boolean} Whether the given input matches the type passed
  */
-export function safeSlot(h, slot) {
-  return slot && slot.length > 1 ? h('div', slot) : slot;
+export function isType(v, type) {
+  return (
+    Object.prototype.toString.call(v).slice(8, -1).toLowerCase() ===
+    type.toLowerCase()
+  );
 }
 
 /**
@@ -35,7 +38,7 @@ export function safeSlot(h, slot) {
  *
  * @param  {HTMLElement} el    [description]
  * @param  {Event}       event [description]
- * @return {undefined}         [description]
+ * @returns {undefined}         [description]
  */
 export function applyFocusTrap(el, event) {
   const focusable = Array.from(el.querySelectorAll(FOCUSABLE));
