@@ -1,8 +1,6 @@
 declare namespace _default {
-    export function bind(el: HTMLElement, binding: import("vue/types/options").DirectiveBinding, vnode: import("vue").VNode, oldVnode: import("vue").VNode): void;
-    export const bind: import("vue").DirectiveFunction;
-    export function unbind(el: HTMLElement, binding: import("vue/types/options").DirectiveBinding, vnode: import("vue").VNode, oldVnode: import("vue").VNode): void;
-    export const unbind: import("vue").DirectiveFunction;
+    const bind: any;
+    const unbind: any;
 }
 export default _default;
 export type ClickoutEl = HTMLElement & {
