@@ -1,4 +1,4 @@
-export { default as autofocus } from "./autofocus";
-export { default as clickout } from "./clickout";
-export { default as copy } from "./copy";
-export { default as intersect } from "./intersect";
+export { default as autofocus } from "./autofocus.js";
+export { default as clickout } from "./clickout.js";
+export { default as copy } from "./copy.js";
+export { default as intersect } from "./intersect.js";
