@@ -1,5 +1,5 @@
 declare namespace _default {
-    export const inserted: import('vue').DirectiveFunction;
+    export const inserted: any;
     export { unbind };
 }
 export default _default;
@@ -8,7 +8,6 @@ export type IntersectEl = HTMLElement & {
 };
 /**
  * @typedef {HTMLElement & { _vtsIntersect?: IntersectionObserver}} IntersectEl
- *
  * @param {IntersectEl} el
  */
 declare function unbind(el: IntersectEl): void;
